fix client: add more field to test.yml, minor refactoring
