+++ client/src/app/sandbox/form-gen-demo/form-gen-demo.component.html
@@ -1,9 +1,11 @@
 <!-- Assuming testForm is your FormGroup -->
 <form (submit)="onSubmit()" [formGroup]="testForm">
     <ng-container *ngFor="let controlName of getControlNames()">
-        <app-input [control]="$any(testForm).get(controlName)" [label]="$any(testForm).get(controlName)?.name"
+        <app-input 
+            [control]="$any(testForm).get(controlName)" 
+            [label]="$any(testForm).get(controlName)?.name"
             [placeholder]="$any(testForm).get(controlName)?.placeholder"
-            [hint]="$any(testForm).get(controlName)?.placeholder"
+            [hint]="$any(testForm).get(controlName)?.hint"
             [inputType]="getInputType($any(testForm).get(controlName)?.type)">
         </app-input>
     </ng-container>
