+++ client/src/assets/test.yml
@@ -2,18 +2,26 @@ parameters:
   - name: РКЧ
     description: Разовая корректировка часов
     type: integer
+    hint: Подсказка для РКЧ
+    placeholder: Placeholder РКЧ
     min: -99
     max: 99
   - name: С_1
     description: Цена импульса датчика объёма
     type: decimal
     precision: 6
+    hint: Подсказка для С_1
+    placeholder: Placeholder С_1
     min: 0
     max: 9.999999
   - name: КВС
     description: Контроль входов ВС
     type: string
+    placeholder: Placeholder КВС
+    hint: Подсказка для КВС
     regex: "[0-3][0-3][0-3]"
   - name: ДО
     description: Дата отсчёта
     type: date
+    hint: Подсказка для ДО
+    placeholder: Placeholder ДО
